home: split the suite at the ceiling, through a harness both halves share
home.test.tsx was 1124 lines, past the 1000 frontend/AGENTS.md sets. It
crossed while Home gained its weekly-review section and then that
section's narrative states.

The seam the file already had is the two weekly describe blocks: one
subject, read against a week of runs rather than against this morning's
queue. What made the split awkward is the harness — Home fans out to
several reads on mount, and stubApi is a recorder rather than a
stand-in, because a staged verdict is only staged if nothing was sent
and no rendering can prove that.

Exporting it from a test file is refused by the lint, rightly. Copying
it would put a SECOND answer to "what does an unrouted read reply with"
in the tree, and these cases turn on exactly that answer. So it is a
module: home.testkit.tsx, named the way recordshell.testkit.tsx is,
because it answers to the app's lint rules rather than a test file's.

Nothing about the tests themselves changed — 33 before, 33 after.

Two things found on the way. The rulebook named three files as being
past the ceiling; there are eighteen, and a list of three reads as "these
are the only ones". It now says the count, and #3232 carries the
enforcement, which needs a ratcheted baseline rather than a flat refusal.

And stubApi's request reader was over the complexity ceiling once it
lived outside a test file, where that rule does not apply. Reading a
call and reading its body are two things now.
