+++ frontend/AGENTS.md
@@ -137,8 +137,8 @@ When writing or touching a screen test:
 - If a test only passes because it got the machine to itself, it is not a test
   yet. Prove it: run the file alone and inside `make fe-unit`, and compare.
 - Test files split at 1000 lines, the same ceiling the Go test trees hold.
-  `onboarding-facts.test.tsx`, `company-act.test.tsx` and
-  `onboarding-conversation.test.tsx` are already past it. Do not grow them.
+  Eighteen are already past it and nothing enforces it yet (#3232). Do not grow
+  one that is over.
 
 ## Storybook is documentation, and it goes stale silently
